pyproject.toml
[project]
name = "ask-acme"
version = "0.1.0"
description = "Production Evals Cookbook - Learn AI evaluation through building Ask Acme"
authors = [{ name = "Ash Tilawat", email = "apply@gauntlethq.com" }]
readme = "README.md"
requires-python = ">=3.11,<3.14"

dependencies = [
    # Core LLM framework
    "langchain>=0.3",
    "langchain-openai>=0.3",
    "langchain-community>=0.3",
    "langgraph>=0.2",

    # Database clients
    "psycopg2-binary>=2.9",
    "pymongo>=4.10",

    # MCP (Model Context Protocol)
    "mcp>=1.0",

    # Embeddings and vector search
    "numpy>=1.26",

    # Data handling
    "pydantic>=2.10",
    "pydantic-settings>=2.7",
    "python-dotenv>=1.0",
    "pyyaml>=6.0",

    # Async support
    "httpx>=0.28",
]

[dependency-groups]
dev = [
    # Notebooks
    "jupyter>=1.0",
    "ipykernel>=6.29",

    # Testing
    "pytest>=8.3",
    "pytest-asyncio>=0.24",

    # Code quality
    "ruff>=0.8",
    "mypy>=1.14",
]

eval = [
    # Evaluation specific
    "pandas>=2.2",
    "tabulate>=0.9",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_ignores = true
ignore_missing_imports = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
